Revert "Feature/GitHub service"
