+++ example_app/build.gradle
@@ -3,13 +3,13 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.jampez.uce_handler"
         minSdkVersion 14
-        targetSdkVersion 30
-        versionCode 2
-        versionName "2.0"
+        targetSdkVersion 29
+        versionCode 1
+        versionName "1.0"
     }
     buildTypes {
         release {
@@ -21,12 +21,11 @@ android {
 
 dependencies {
     api fileTree(include: ['*.jar'], dir: 'libs')
-    api 'androidx.appcompat:appcompat:1.3.0-alpha01'
+    api 'androidx.appcompat:appcompat:1.2.0-beta01'
     api project(':uce_handler')
     api project(':ignore-dummy-module')
-    implementation "androidx.core:core-ktx:1.3.0"
+    implementation "androidx.core:core-ktx:1.2.0"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 }
 repositories {
     mavenCentral()
