+++ example_app/src/main/res/values/strings.xml
@@ -1,7 +1,4 @@
 <resources>
-  <string name="app_name">UCE-Handler</string>
-  <string name="crash_the_app">Crash This App</string>
-
-  <!-- TODO: Remove or change this placeholder text -->
-  <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="app_name">UCE-Handler</string>
+    <string name="crash_the_app">Crash This App</string>
 </resources>
