+++ uce_handler/build.gradle
@@ -3,11 +3,11 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 29
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 30
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -18,38 +18,13 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
 }
-
-def libraries = [
-        koinVersion             : "2.0.1"
-]
 
 dependencies {
     api fileTree(dir: 'libs', include: ['*.jar'])
-    api 'androidx.appcompat:appcompat:1.3.0-alpha01'
-
-    api 'com.squareup.retrofit2:retrofit:2.6.2'
-    api 'com.squareup.retrofit2:converter-gson:2.6.2'
-    api 'com.squareup.okhttp3:logging-interceptor:4.7.2'
-
-    api "androidx.core:core-ktx:1.3.0"
-    api "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-
-    // Koin
-    api "org.koin:koin-android:${libraries.koinVersion}"
-    api "org.koin:koin-androidx-scope:${libraries.koinVersion}"
-    api "org.koin:koin-androidx-viewmodel:${libraries.koinVersion}"
-
-    api 'androidx.navigation:navigation-fragment-ktx:2.3.0'
-    api 'androidx.navigation:navigation-ui-ktx:2.3.0'
-    
-    api 'androidx.constraintlayout:constraintlayout:1.1.3'
-
-    api 'com.google.android.material:material:1.1.0'
+    api 'androidx.appcompat:appcompat:1.2.0-beta01'
+    implementation "androidx.core:core-ktx:1.2.0"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 }
 repositories {
     mavenCentral()
