+++ uce_handler/src/main/java/com/jampez/uceh/UCEFileProvider.kt
@@ -1,0 +1,5 @@
+package com.jampez.uceh
+
+import androidx.core.content.FileProvider
+
+class UCEFileProvider : FileProvider()
