--- uce_handler/src/main/java/com/jampez/uceh/features/bitbucket/BitBucketPostData.kt
@@ -1,11 +1,0 @@
-package com.jampez.uceh.features.bitbucket
-
-import com.google.gson.annotations.SerializedName
-
-data class BitBucketPostData(
-        @SerializedName("title")
-        val title: String,
-
-        @SerializedName("content")
-        val content: Content
-)
