--- uce_handler/src/main/java/com/jampez/uceh/features/bitbucket/BitBucketResponse.kt
@@ -1,9 +1,0 @@
-package com.jampez.uceh.features.bitbucket
-
-import com.google.gson.annotations.SerializedName
-
-data class BitBucketResponse(
-        @SerializedName("id") val id: Long,
-        @SerializedName("title") val title: String,
-        @SerializedName("type") val type: String
-)
