--- uce_handler/src/main/java/com/jampez/uceh/features/bitbucket/Content.kt
@@ -1,8 +1,0 @@
-package com.jampez.uceh.features.bitbucket
-
-import com.google.gson.annotations.SerializedName
-
-data class Content(
-        @SerializedName("raw")
-        val raw: String?
-)
