--- uce_handler/src/main/java/com/jampez/uceh/features/gitlab/GitLabPostData.kt
@@ -1,14 +1,0 @@
-package com.jampez.uceh.features.gitlab
-
-import com.google.gson.annotations.SerializedName
-
-data class GitLabPostData(
-        @SerializedName("title")
-        val title: String,
-
-        @SerializedName("description")
-        val description: String?,
-
-        @SerializedName("labels")
-        val labels: List<String>
-)
