--- uce_handler/src/main/java/com/jampez/uceh/features/network/RetrofitClient.kt
@@ -1,45 +1,0 @@
-package com.jampez.uceh.features.network
-
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-object RetrofitClient {
-    var retrofit: Retrofit? = null
-
-    fun getClient(baseUrl: String): Retrofit? {
-        if (retrofit == null) {
-            //TODO While release in Google Play Change the Level to NONE
-            val interceptor = HttpLoggingInterceptor()
-            interceptor.level = HttpLoggingInterceptor.Level.BODY
-            val client = OkHttpClient.Builder()
-                    .addInterceptor(interceptor)
-                    .connectTimeout(100, TimeUnit.SECONDS)
-                    .readTimeout(100, TimeUnit.SECONDS)
-                    .build()
-
-            retrofit = Retrofit.Builder()
-                    .client(client)
-                    .baseUrl(baseUrl)
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .build()
-        }
-
-        return retrofit
-
-    }
-
-    private fun authenticationInterceptor(token: String) = Interceptor { chain ->
-        if (token.isNotEmpty()) {
-            chain.proceed(chain.request().newBuilder()
-                    .addHeader("Authorization", "Token $token")
-                    .build())
-        } else {
-            chain.proceed(chain.request().newBuilder()
-                    .build())
-        }
-    }
-}
