--- uce_handler/src/main/java/com/jampez/uceh/features/supportissue/SupportIssueModule.kt
@@ -1,5 +1,0 @@
-package com.jampez.uceh.features.supportissue
-
-import org.koin.dsl.module
-
-val supportIssueModule = module { single { SupportIssueRepository() } }
