--- uce_handler/src/main/java/com/jampez/uceh/features/uce/UCEFileProvider.kt
@@ -1,5 +1,0 @@
-package com.jampez.uceh.features.uce
-
-import androidx.core.content.FileProvider
-
-class UCEFileProvider : FileProvider()
