--- uce_handler/src/main/java/com/jampez/uceh/utils/AppException.kt
@@ -1,3 +1,0 @@
-package com.jampez.uceh.utils
-
-class AppException(val exception: Throwable?) : Exception()
