--- uce_handler/src/main/res/layout/fragment_main.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  android:id="@+id/main_layout"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools">
-
-  <com.google.android.material.textview.MaterialTextView
-    android:id="@+id/title"
-    android:textStyle="bold"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:padding="15dp"
-    android:gravity="center"
-    android:text="@string/something_went_wrong"
-    android:textSize="25sp"
-    app:layout_constraintLeft_toLeftOf="parent"
-    app:layout_constraintRight_toRightOf="parent"
-    app:layout_constraintTop_toTopOf="parent" />
-
-  <com.google.android.material.textview.MaterialTextView
-    android:id="@+id/subtitle"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:gravity="center"
-    android:textSize="18sp"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/title"
-    android:text="@string/ask_for_error_log" />
-
-  <ImageView
-    android:id="@+id/background_image"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:padding="15dp"
-    android:contentDescription="@string/app_name"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/subtitle"
-    tools:src="@drawable/crash_bg" />
-
-  <RelativeLayout
-    android:id="@+id/support_ticket_layout"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:layout_constraintLeft_toLeftOf="parent"
-    app:layout_constraintRight_toRightOf="parent"
-    app:layout_constraintTop_toBottomOf="@id/background_image"
-    app:layout_constraintBottom_toTopOf="@id/button_view_error_log">
-
-    <androidx.core.widget.ContentLoadingProgressBar
-      android:id="@+id/support_ticket_request_progress"
-      android:layout_centerInParent="true"
-      style="?android:attr/progressBarStyleLarge"
-      android:layout_width="wrap_content"
-      android:minWidth="100dp"
-      android:layout_height="wrap_content"
-      android:minHeight="100dp"/>
-
-    <Button
-      android:layout_centerInParent="true"
-      style="?attr/materialButtonStyle"
-      android:id="@+id/button_create_support_ticket"
-      android:visibility="invisible"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:text="@string/create_support_ticket"/>
-
-    <LinearLayout
-      android:id="@+id/support_ticket_text_layout"
-      android:orientation="vertical"
-      android:visibility="invisible"
-      android:layout_centerInParent="true"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-
-      <TextView
-        android:id="@+id/support_ticket_response"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:textSize="18sp"
-        android:textColor="@color/colorPrimaryDark"
-        tools:text="@string/ask_for_error_log" />
-
-      <TextView
-        android:id="@+id/support_ticket_number"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:layout_marginTop="10dp"
-        android:textSize="18sp"
-        android:textColor="@color/colorPrimaryDark"
-        tools:text="@string/ticket_number" />
-
-    </LinearLayout>
-
-  </RelativeLayout>
-
-
-  <androidx.appcompat.widget.AppCompatButton
-    android:id="@+id/button_view_error_log"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:text="@string/view_error_log"
-    android:visibility="gone"
-    style="@style/Widget.AppCompat.Button.Borderless"
-    android:textAllCaps="false"
-    android:background="@android:color/transparent"
-    android:textColor="@color/colorPrimaryDark"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintStart_toStartOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
