--- uce_handler/src/main/res/navigation/nav_graph.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:id="@+id/nav_graph.xml"
-  app:startDestination="@id/mainFragment">
-
-  <fragment
-    android:id="@+id/mainFragment"
-    android:name="com.jampez.uceh.features.uce.MainFragment"
-    tools:layout="@layout/fragment_main"
-    android:label="MainFragment" >
-
-  </fragment>
-
-</navigation>
