+++ uce_handler/src/main/res/values/colors.xml
@@ -2,8 +2,4 @@
 <resources>
     <color name="white">#FFFFFF</color>
     <color name="black">#000000</color>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#999999</color>
-    <color name="background">#FAFAFA</color>
 </resources>
